backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
torch==2.7.0
torchvision==0.22.0
pillow==10.4.0
tqdm==4.66.5
